refactor: update styles
