+++ train-a/src/app/orders/pages/orders-page/orders.component.html
@@ -3,7 +3,7 @@
   <span *ngIf="isManager">Orders</span>
 </mat-toolbar>
 <div class="spinner" *ngIf="isLoading$ | async">
-  <mat-spinner strokeWidth="3"> </mat-spinner>
+  <mat-spinner strokeWidth="3"></mat-spinner>
 </div>
 <div class="orders-container">
   <app-order-card *ngFor="let order of sortedOrders" [order]="order"></app-order-card>
