+++ train-a/src/app/orders/pages/orders-page/orders.component.scss
@@ -14,6 +14,9 @@
 }
 
 .spinner {
+  position: fixed;
+  z-index: 1000;
+  inset: 0;
   display: flex;
   align-items: center;
   justify-content: center;
