+++ train-a/src/app/shared/services/api.service.ts
@@ -2,7 +2,7 @@ import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 
-import { Order } from '../../orders/state/order.model';
+import { Order } from '../../orders/models/order.model';
 
 @Injectable({
   providedIn: 'root',
